requirements.txt
langchain
python-dotenv
langchain-community
streamlit
langchain-groq
chromadb
sentence-transformers
pypdf